requirements.txt
# FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# LLM & Agent
openai>=1.12.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0

# Database
supabase>=2.9.0

# Utilities
python-dotenv>=1.0.0
tiktoken>=0.7.0
requests>=2.31.0

# Media Processing
pillow>=10.0.0
opencv-python-headless>=4.8.0
moviepy>=1.0.3
imageio-ffmpeg>=0.4.9
